requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
ollama==0.4.2
pyaml==24.9.0
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyYAML==6.0.2
rich==13.9.4
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
